requirements/dev.txt
-r test.txt
git+https://github.com/tsileo/little-boxes.git
html2text
pyyaml
flake8
ipython
ipdb